+++ .gitignore
@@ -294,3 +294,4 @@ MvcWebsite.v12.suo
 MvcWebsite.v12.suo
 MvcWebsite/MvcWebsite.csproj
 *.suo
+MvcWebsite.v12.suo
