+++ MvcWebsite/MessageBroker/IMessageBrokerApi.cs
@@ -11,7 +11,6 @@ namespace MvcWebsite.MessageBroker
 {
     public interface IMessageBrokerApi
     {
-        HttpClient CreateClient();
         IEnumerable<CommentModel> GetPageComments(string pageFilter);
         IEnumerable<CommentModel> GetComments();
         CommentModel GetComment(int id);
